fix: visual regressions
